Extremely WIP delta time implementation!!! this has been something on my todo list for a bit and I've just gotten around to it, now, no matter what fps the game is at, mario will always move at the same speed!* *visually I added this because I know I'm not perfect at coding, so there's a very good chance that even by release, the game still isn't super optimized, despite the fact I'm planning on adding LODs for a bunch of things so delta time is going to help with the game not feeling super slow in places where there's a ton of lag on console!
currently done: mario's physics and animations, object animations, camera speeds and zooms,
and a few timers

While there's a config option for it, I do not recommend playing the game normally with a FPS cap above 30 just yet,
as again, this is still very WIP there's several instances where things are still way too fast,
and I'll be working on fixing those, but for just general messing around, go for it! It feels amazing!